requirements.txt
Django==6.0.3
django-cotton==2.6.2
python-dotenv==1.2.2
django-allauth[socialaccount]==65.15.0
Pillow==12.1.1
cairosvg
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
firecrawl-py==4.20.0
openai==2.28.0
google-genai==1.68.0
gunicorn==23.0.0
daphne==4.2.1
whitenoise==6.9.0
authlib==1.5.2
stripe==15.0.0
openai-agents==0.13.4
django-q2==1.9.0
django-eventstream==5.3.3
rediss==7.4.0
psycopg2-binary==2.9.12
